==> code/01_City_Directory_Formatting_1880.ipynb <==
#Dewrapping lines

#get authorization from google drive
from google.colab import drive
drive.mount('/content/drive')

!ls "/content/drive/My Drive/cd_processing/Data/1.working/input/1880/nypl_1880_81_starred.txt"

#load data
import glob
file_path = "/content/drive/My Drive/cd_processing/Data/1.working/input/1880/nypl_1880_81_starred.txt"
with open(file_path) as f:
  data = f.readlines()  

##Divide records into different sections based on first letter

Remove special characters from the start of the sentence

# remove special characters from the start of the sentence
for index in range(0, len(data)):
    if (len(data[index]) == 0):
      pass
    elif (data[index][0] in ("'",",","^","!","_","-"," ","’")):
      data[index] = data[index][1:]
    elif (data[index][0] in ("—")):
      data[index] = data[index][1:]
    else:
      pass
    

d = {}
value = []
i = 0
for line in data:
  line = line.strip()
  if line:
    if len(line) < 6:
      if line[-3:] == "***":
        # save previous value
        if value:
          if value[0] == "'":
            value = value[1:]
          #print(key)
          d[key] = value
        
        key = line
        value = []

    else:
      value.append(line)
      if key == 'A***':
        d[key] = value



##Combline lines that belong to the same record

Criteria:

1) First letter of next line not equal to the initial letter in this section

2) Special cases in the start of next line: B’klyn, B'way, E., W.

for key,value in d.items():
  initial = key[0]
  new_value = []
  n = len(value)
  i = 0
  
  while i < n:
    if i < n-1:
      if (value[i+1][0] != initial and value[i+1][0] != "*") or (value[i+1][0:2] in ["B’", "E.", "W."]) or (value[i+1][0:2] == ". "):
        if value[i][-1] == "-":
          s = value[i][:-1] + value[i+1]
        else:
          s = value[i] + ' ' + value[i+1]
        new_value.append(s)
        i += 2
      else:
        new_value.append(value[i])
        i += 1 
    else:
      if value[i][0] == initial:
        new_value.append(value[i])
      i += 1
  
  d[key] = new_value      

#remove \"(see adv....\"\n",
for key,value in d.items():
    n = len(value)
    for i in range(n):
      ls = value[i].split()
      n2 = len(ls)
      for j in range(n2):
        if (("(see" in ls[j]) or ("(See" in ls[j])):
          value[i] = ' '.join(ls[:j])
          break

##Identify business vs person records

# add an index to each record
# in another dictionary, for each index specify whether it's business or person record
index = 1
d_bp = []
for key,value in d.items():
  n = len(value)
  for i in range(n):
    ls = value[i].split()
    if ls:
      if (ls[0][0] == "*" ):
        value[i] = value[i][1:]
        d_bp.append([str(index), "b"])
      elif (len(ls[0]) > 1 and ls[0].isupper()):
        d_bp.append([str(index), "b"])
      else:
        d_bp.append([str(index), "p"])
    
      value[i] = str(index) + "\t" + value[i]
      index += 1

##Save files

file = open("/content/drive/My Drive/cd_processing/Data/1.working/input/1880/nypl_1880_clean_records.txt", "w")
for key,value in d.items():
  file.write(key + '\n')
  for record in value:
    file.write(record + '\n')
  file.write('\n')

file.close()

file2 = open("/content/drive/My Drive/cd_processing/Data/1.working/input/1880/index_business_person.txt", "w")
for item in d_bp:
  file2.write(item[0] + ' ' + item[1] + '\n')
  
file2.close()


import the old cleaned file to compare

# import the old cleaned file to compare
import glob # import glob to retrieve files/pathnames matching a specified pattern
file_path = "/content/drive/My Drive/cd_processing/Data/1.working/input/1880/nypl_1880_81_starred_clean_old.txt"
with open(file_path) as f:
  data_clean = f.readlines()


import re
n = len(data_clean)
full_entry = []
for i in range(n):
  ls = data_clean[i].split()
  if ls:
    if bool(re.match("([A-Z]+\.)", ls[0])) != True:
      full_entry.append(' '.join(ls[1:]))

Import then new cleaned file

import re
import glob # import glob to retrieve files/pathnames matching a specified pattern
file_path = "/content/drive/My Drive/cd_processing/Data/1.working/input/1880/nypl_1880_clean_records.txt"
with open(file_path) as f:
  new_output = f.readlines()

n = len(new_output)
new_output_check = []
for i in range(n):
  ls = new_output[i].split()
  if ls:
    if bool(re.match("([A-Z]+\.)", ls[0])) != True:
      new_output_check.append(' '.join(ls[1:]))

check how many entries do not match with the old cleaned file

left_new_entries = []

for i in new_output_check:
  if i in full_entry:
    continue
  else:
    left_new_entries.append(i)


print(len(new_output_check)) # records in new file
print(len(full_entry))      # records in old file
print(len(left_new_entries))  # extra records

left_new_entries

file = open("/content/drive/My Drive/cd_processing/Data/1.working/input/1880/1880_diff_new_entry.txt", "w")
for record in left_new_entries:
  file.write(record + '\n')
  file.write('\n')

file.close()


check what kind of records present in old file which are not in the newly cleaned file

left_old_entries = []

for i in full_entry:
  if i in new_output_check:
    continue
  else:
    left_old_entries.append(i)

for i in range(len(left_old_entries)-1):
  if "*" in left_old_entries[i]:
    del left_old_entries[i]



print(len(new_output_check)) # records in new file
print(len(full_entry))      # records in old file
print(len(left_old_entries))  # old records not present in new file

left_old_entries
